move SQLSegmentAssert to root package
